api/django_project/pyproject.toml
[tool.poetry]
name = "django_project"
version = "0.1.0"
description = ""
authors = ["John Doherty <john.doherty@paragon-cloud.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^23.2"
celery = "^5.3.0"
django-celery-beat = "^2.5.0"
channels = "^3.0.5"
redis = "^4.3.4"
flower = "^1.2.0"
django-cors-headers = "^3.13.0"
channels-redis = "^3.4.1"
SQLAlchemy = "^1.4.39"
auth0-python = "^3.23.1"
cryptography = "^37.0.4"
drf-jwt = "^1.19.2"
PyJWT = "^2.4.0"
requests = "^2.28.1"
elasticsearch = "^8.5.0"
gunicorn = "^20.1.0"
celery-singleton = "^0.3.1"
flake8 = "^6.0.0"
mypy = "^0.991"
black = "^22.12.0"
isort = "^5.10.1"
pylint = "^2.15.8"
autopep8 = "^2.0.0"
types-requests = "^2.28.11"
whitenoise = "^6.2.0"
pydantic = "^1.10.8"
injector = "^0.20.1"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
debugpy = "^1.6.7"
geopy = "^2.3.0"
boto3 = "^1.26.162"
sendgrid = "^6.10.0"
transitions = "^0.9.0"
redis-lock = "^0.2.0"
types-redis = "^4.6.0.2"
faker = "^19.2.0"
types-python-dateutil = "^2.8.19.14"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-exporter-otlp = "^1.20.0"
opentelemetry-instrumentation-django = "^0.41b0"
opentelemetry-instrumentation-celery = "^0.41b0"
opentelemetry-instrumentation-elasticsearch = "^0.41b0"
ics = "^0.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 79
ignore = "E501,W6"
in-place = true
recursive = true
aggressive = 3

[tool.isort]
profile = "black"